use stdint.h

+++ c/linux/Platform.h
@@ -1,12 +1,13 @@
 #ifndef PLATFORM_H
 #define PLATFORM_H
 
+//
 class Platform {
 public:
-  int refresh_time;
+  uint16_t refresh_time;
   double loadavg[3];
   utsname info;
-  unsigned long uptime, idle;
+  uint64_t uptime, idle;
 
 public:
   Platform();
